move files as were in main
