+++ src/data/network/graph/implementation/theGraphAPIClientMapper.js
@@ -2,11 +2,11 @@
 /* eslint-disable max-lines-per-function */
 
 import GraphQLAPIMapper from "../graphQLAPIMapper"
-import Thread from "@/models/marketplace/thread"
-import Needle from "@/models/marketplace/needle"
+import Thread from "@/models/thread"
+import Needle from "@/models/needle"
 import Erc20 from "@/models/erc20"
-import { Vote } from "../../../../models/vote"
-import { MarketOrder } from "../../../../models/marketplace/marketOrder"
+import { Vote } from "@/models/vote"
+import { MarketOrder } from "@/models/marketOrder"
 import {
   mapPaperProposals,
   mapUpgradeProposals,
